requirements.txt
langchain==0.0.350
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=3.19.0
streamlit
google-generativeai
python-dotenv
# 0.0.284